broke apart ai.ts file into subfiles
